+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Development
 
-
+- Replace IRoutes interface with IBlueprint.
 
 ## [0.3.0](https://github.com/berlinonline/ckanext-dcatde_berlin/releases/tag/v0.3.0)